feat(player): player controls, and fix renderer

+++ include/renderer.h
@@ -4,6 +4,6 @@
 typedef struct Renderer Renderer;
 
 Renderer* create_renderer(int screen_width, int screen_height, int renderer_width, int renderer_height);
-void renderer_render(Renderer* renderer, int camera_pos_x, int camera_pos_y, double camera_angle);
+void renderer_render(Renderer* renderer, double camera_pos_x, double camera_pos_y, double camera_angle);
 
 #endif
